+++ src/client/components/SignInModal.spec.js
@@ -29,9 +29,9 @@ describe('<SignInModal />', () => {
   // close modal by clicking cancel
   // close modal by clicking outside of modal
 
-  it('should contain <form />', () => {
+  it('should contain <input />', () => {
     const { wrapper } = setup();
-    expect(wrapper.find('form'))
+    expect(wrapper.find('input'))
       .toHaveLength(1);
   });
