+++ src/client/components/chatmain/ChatMain.jsx
@@ -13,7 +13,7 @@ import ChatHeader      from './ChatHeader.jsx';
 import ChatToolbar     from './ChatToolbar.jsx';
 import ChatContent     from './ChatContent.jsx';
 import ChatInput       from './ChatInput.jsx';
-import SignInModal     from './SignInModal.jsx';
+import SignInModal     from '../SignInModal.jsx';
 import                      './chatmain.scss';
 
 export class ChatMain extends React.Component {
