+++ src/client/components/chatmain/chatmain.scss
@@ -36,6 +36,7 @@
 .chat-container {
   display: flex;
   flex-direction: row;
+  overflow: hidden;
 }
 
 .chat-header-wrapper {
@@ -64,7 +65,7 @@
   flex-shrink: 0;
   display: flex;
   width: 70px;
-  // 
+  //
   border-left: 1px solid #f1f1f1;
 }
 
@@ -167,7 +168,7 @@
   }
 
   .chat-item-details {
-    // 
+    //
   }
 
   .chat-item-user {
@@ -417,4 +418,23 @@
   -webkit-flex-shrink: 0;
   -ms-flex-negative: 0;
   flex-shrink: 0;
+}
+
+.modal-footer-btn {
+  margin: 0;
+  font-size: 12px;
+  font-weight: 400;
+  line-height: 19px;
+  box-sizing: border-box;
+  display: inline-block;
+  padding: 9.6px 42px;
+  background-color: #46bc99;
+  border: none;
+  border-radius: 6px;
+  outline: none;
+  color: #ffffff;
+  text-transform: uppercase;
+  letter-spacing: 0.1em;
+  text-align: center;
+  transition: background-color 0.1s ease, border-color 0.1s ease;
 }
